🚚 Refactor(InputFloat, InputBasic): A strategic relocation is carried out for these 2 components. 🚚
